move all schemas to core app
